api/requirements.txt
fastapi
uvicorn
pymongo
pillow
pyzbar
python-multipart
